+++ src/includes/classes/SCore/Utils/SysOption.php
@@ -43,15 +43,14 @@ class SysOption extends Classes\SCore\Base\Core
         $key = $this->App->Config->©brand['©var'].'_'.$key;
 
         if ($this->App->Config->§specs['§is_network_wide'] && $this->Wp->is_multisite) {
-            if (isset($value)) {
-                update_network_option(null, $key, $value);
+            if (isset($value)) { // Updating?
+                update_network_option(null, $key, $value); // Autoload not possible.
             }
             if (($value = get_network_option(null, $key)) === null || $value === false) {
-                add_network_option(null, $key, ':null'); // Autoload impossible.
-                // These will not autoload and there is no way to change this.
+                add_network_option(null, $key, ':null'); // Autoload not possible.
             }
-        } else { // Default.
-            if (isset($value)) {
+        } else { // Default behavior.
+            if (isset($value)) { // Updating?
                 update_option($key, $value, $autoload ? 'yes' : 'no');
             }
             if (($value = get_option($key)) === null || $value === false) {